+++ src/Managers/TableSettingsManager.php
@@ -3,23 +3,42 @@
 namespace Glorand\Model\Settings\Managers;
 
 use Glorand\Model\Settings\Contracts\SettingsManagerContract;
+use Glorand\Model\Settings\Exceptions\ModelSettingsException;
 use Glorand\Model\Settings\Models\ModelSettings;
+use Glorand\Model\Settings\Traits\HasSettings;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Cache;
 
 /**
- * Class TableSettingsManager
- * @package Glorand\Model\Settings\Managers
- * @property  \Illuminate\Database\Eloquent\Model|\Glorand\Model\Settings\Traits\HasSettingsTable $model
+ * @property Model|HasSettings $model
  */
 class TableSettingsManager extends AbstractSettingsManager
 {
     /**
-     * @param  array  $settings
-     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
-     * @throws \Exception
+     * @throws ModelSettingsException
+     */
+    public function getStoredValue(): array
+    {
+        $this->ensureModelIsPersisted();
+
+        if (config('model_settings.drivers.table.use_cache', true)) {
+            return Cache::rememberForever($this->model->getSettingsCacheKey(), function () {
+                return $this->fetchStoredValue();
+            });
+        }
+
+        return $this->fetchStoredValue();
+    }
+
+    /**
+     * @param array $settings
+     * @return SettingsManagerContract
+     * @throws ModelSettingsException
      * @SuppressWarnings(PHPMD.ElseExpression)
      */
     public function apply(array $settings = []): SettingsManagerContract
     {
+        $this->ensureModelIsPersisted();
         $this->validate($settings);
 
         $modelSettings = $this->model->modelSettings()->first();
@@ -40,5 +59,14 @@ class TableSettingsManager extends AbstractSettingsManager
         cache()->forget($this->model->getSettingsCacheKey());
 
         return $this;
+    }
+
+    private function fetchStoredValue(): array
+    {
+        if ($modelSettings = $this->model->modelSettings()->first()) {
+            return $modelSettings->settings;
+        }
+
+        return [];
     }
 }