+++ tests/Models/UserWithDeclaredDriver.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Glorand\Model\Settings\Tests\Models;
+
+/**
+ * Class UserWithDeclaredDriver
+ * @package Glorand\Model\Settings\Tests\Models
+ */
+class UserWithDeclaredDriver extends UserWithField
+{
+    protected $settingsDriver = 'redis';
+}