+++ tests/Models/UsersWithTable.php
@@ -3,20 +3,20 @@
 namespace Glorand\Model\Settings\Tests\Models;
 
 use Glorand\Model\Settings\Contracts\SettingsManagerContract;
-use Glorand\Model\Settings\Traits\HasSettingsTable;
+use Glorand\Model\Settings\Traits\HasSettings;
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * Class UsersWithTable
- * @package Glorand\Model\Settings\Tests\Models
  * @method static first()
  * @method SettingsManagerContract config()
  */
 class UsersWithTable extends Model
 {
-    use HasSettingsTable;
+    use HasSettings;
 
-    public $invokeSettingsBy = 'config';
+    protected $settingsDriver = 'table';
+
+    public string $invokeSettingsBy = 'config';
 
     protected $table = 'users_with_table';